Fix SDG story page issue
